Fixing critical inexing bugs in Radon transform code.
Both bugs are related to secitoning the image into 'windows.' The first caused the code to occasionally calculate an invalid image index for some combinations of window size and window step. The other was related to windows that span consecutive frames. The desired behavior is to append the end of the initial frame to the beginning of the subsequent frame to create an appropriately sized window, however, the code was reading the same image frame twice and appending the end of that frame to the beginning of that same frame, generating a composite window of non-sequential lines. It should be noted that the window size also is 1 pixel larger than requested. This will be fixed eventually but is not fixed here.